+++ app/src/main/java/com/juan/notes/ui/notesDetail/NoteDetailViewModel.kt
@@ -23,7 +23,7 @@ class NoteDetailViewModel @Inject constructor(
         }
     }
 
-    fun getNote(id: Int): LiveData<Note> {
+    fun getNote(id: Long): LiveData<Note> {
         return getNoteByIdUseCase(id)
     }
 }
